refactor: extract homepage to pure react
